Added the Redis DB for the expired token handling

+++ models/item.py
@@ -6,7 +6,7 @@ class ItemModel(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(80), unique=False, nullable=False)
-    price = db.Column(db.Float(precision=2), unique=False, nullable=False)
+    price = db.Column(db.Numeric(10, 2), unique=False, nullable=False)
 
     description = db.Column(db.String)
